+++ UpToolLib/Tool/AppInstall.cs
@@ -14,8 +14,8 @@ namespace UpToolLib.Tool
         /// Install an application
         /// </summary>
         /// <param name="appI">The app to install</param>
-        /// <param name="download">A method to download files. Input: app file, Outputs: whether the download was successful and the data</param>
-        public static void Install(App appI)
+        /// <param name="force">Set to true to overwrite all old data</param>
+        public static void Install(App appI, bool force)
         {
             string app = "";
             string tmp = "";
@@ -28,9 +28,17 @@ namespace UpToolLib.Tool
                 if (Directory.Exists(tmp))
                     Directory.Delete(tmp, true);
                 Directory.CreateDirectory(tmp);
-                if (Directory.Exists(app))
-                    Directory.Delete(app, true);
-                Directory.CreateDirectory(app);
+                if (force)
+                {
+                    if (Directory.Exists(app))
+                        Directory.Delete(app, true);
+                    Directory.CreateDirectory(app);
+                }
+                else
+                {
+                    if (!Directory.Exists(app))
+                        Directory.CreateDirectory(app);
+                }
                 (bool dlSuccess, byte[] dlData) = ExternalFunctionalityManager.instance.Download(new Uri(appI.File));
                 if (!dlSuccess)
                     throw new Exception("Download failed");
@@ -44,7 +52,7 @@ Package: {pkgHash}
 Online: {appI.Hash.ToUpper()}");
                 }
                 File.WriteAllBytes(Path.Combine(app, "package.zip"), dlData);
-                CompleteInstall(appI);
+                CompleteInstall(appI, force);
 #if !DEBUG
             }
             catch
@@ -63,7 +71,7 @@ Online: {appI.Hash.ToUpper()}");
 #endif
         }
 
-        public static void InstallZip(string zipPath, App meta)
+        public static void InstallZip(string zipPath, App meta, bool force)
         {
             string app = "";
             string tmp = "";
@@ -76,7 +84,7 @@ Online: {appI.Hash.ToUpper()}");
                     Directory.Delete(tmp, true);
                 Directory.CreateDirectory(tmp);
                 File.Copy(zipPath, Path.Combine(app, "package.zip"));
-                CompleteInstall(meta);
+                CompleteInstall(meta, force);
             }
             catch
             {
@@ -94,13 +102,19 @@ Online: {appI.Hash.ToUpper()}");
         //Use
         //PowerShell -Command "Add-Type -AssemblyName PresentationFramework;[System.Windows.MessageBox]::Show('Hello World')"
         //for message boxes
-        private static void CompleteInstall(App app) => CompleteInstall(app.appPath, app.Name, app.Description, app.Version, app.MainFile);
+        private static void CompleteInstall(App app, bool force) => CompleteInstall(app.appPath, app.Name, app.Description, app.Version, app.MainFile, force);
 
-        private static void CompleteInstall(string appPath, string name, string description, Version version, string mainFile)
+        private static void CompleteInstall(string appPath, string name, string description, Version version, string mainFile, bool force)
         {
             string tmp = PathTool.tempPath;
             ZipFile.ExtractToDirectory(Path.Combine(appPath, "package.zip"), tmp);
-            Directory.Move(Path.Combine(tmp, "Data"), Path.Combine(appPath, "app"));
+            if (force)
+                Directory.Move(Path.Combine(tmp, "Data"), Path.Combine(appPath, "app"));
+            else
+            {
+                CopyAll(Path.Combine(tmp, "Data"), Path.Combine(appPath, "app", "Data"));
+                Directory.Delete(Path.Combine(tmp, "Data"), true);
+            }
             XElement el = new XElement("app", new XElement("Name", name), new XElement("Description", description),
                 new XElement("Version", version));
             if (mainFile != null)
@@ -114,6 +128,17 @@ Online: {appI.Hash.ToUpper()}");
                 CreateNoWindow = true,
                 WindowStyle = ProcessWindowStyle.Hidden
             }).WaitForExit();
+        }
+        
+        private static void CopyAll(string source, string target)
+        {
+            if (string.Equals(Path.GetFullPath(source), Path.GetFullPath(target), StringComparison.CurrentCultureIgnoreCase))
+                return;
+            if (!Directory.Exists(target))
+                Directory.CreateDirectory(target);
+            foreach (string file in Directory.GetFiles(source)) File.Copy(file, Path.Combine(target, Path.GetFileName(file)), true);
+            foreach (string dir in Directory.GetDirectories(source))
+                CopyAll(dir, Path.Combine(target, Path.GetFileName(dir)));
         }
     }
 }
